# api.yaml
openapi: 3.0.3
info:
  title: Gestión de seguimiento docente
  description: |-
    API para gestionar el seguimiento docente
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0
servers:
  - url: http://localhost:8070/api/v1/
...
paths:
...
  /espacios/actividades/{idActividad}:
    get:
      tags:
        - espacios
        - actividades
      summary: Devuelve una lista de los espacios asociados a una actividad
      description: Devuelve una lista de espacios relacionados a la actividad con id = {idActividad}
      operationId: getEspaciosOfActividad
      parameters:
        - $ref: '#/components/parameters/idActividad'
      responses:
        '200':
          $ref: '#/components/responses/EspacioListaData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Actividad no encontrada
...
  /actividades/usuarios/{idUsuario}:
    get:
      tags:
        - actividades
        - usuarios
      summary: Devuelve una lista de las actividades de un usuario por su id
      description: Devuelve una lista de las actividades del usuario con id = {idUsuario}.
      operationId: getActividadesOfUsuario
      parameters:
        - $ref: '#/components/parameters/idUsuario'
      responses:
        '200':
          $ref: '#/components/responses/ActividadListaData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Usuario no encontrado
...
  /actividades/{idActividad}:
    get:
      tags:
        - actividades
      summary: Devuelve información sobre una actividad por su id
      description: Devuelve parámetros informativos de la actividad con id = {idActividad}
      operationId: getActividadById
      parameters:
        - $ref: '#/components/parameters/idActividad'
      responses:
        '200':
          $ref: '#/components/responses/Actividad'
        '400':
          description: Id suministrado no válido
        '404':
          description: Actividad no encontrada
  /clases/{idClase}:
    get:
      tags:
        - clases
      summary: Devuelve información sobre una clase por su id
      description: Devuelve la asignatura y el grupo al que hace referencia la clase con id = {idClase}
      operationId: getClaseById
      parameters:
        - $ref: '#/components/parameters/idClase'
      responses:
        '200':
          $ref: '#/components/responses/ClaseData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Clase no encontrada
...
  /asignaturas/{idAsignatura}:
    get:
      tags:
        - asignaturas
      summary: Devuelve información de una asignatura por su id
      description: Devuelve información de la asignatura con id = {idAsignatura}
      operationId: getAsignaturaById
      parameters:
        - $ref: '#/components/parameters/idAsignatura'
      responses:
        '200':
          $ref: '#/components/responses/AsignaturaData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Asignatura no encontrada
  /grupos/{idGrupo}:
    get:
      tags:
        - grupos
      summary: Devuelve información de un grupo por su id
      description: Devuelve información del grupo con id = {idAsignatura}
      operationId: getGrupoById
      parameters:
        - $ref: '#/components/parameters/idGrupo'
      responses:
        '200':
          $ref: '#/components/responses/GrupoData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Grupo no encontrado
...
  /excepciones/{idExcepcion}:
    get:
      tags:
        - excepciones
      summary: Devuelve información de una excepción por su id
      description: Devuelve información sobre la excepción con id = {idExcepcion}
      operationId: getExcepcionById
      parameters:
        - $ref: '#/components/parameters/idExcepcion'
      responses:
        '200':
          $ref: '#/components/responses/ExcepcionData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Excepcion no encontrada
  /excepciones/actividades/{idActividad}:
    get:
      tags:
        - excepciones
        - actividades
      summary: Devuelve una lista de excepciones de una actividad por su id
      description: Devuelve la lista de excepciones asociadas a la actividad con id = {idActividad}
      operationId: getExcepcionesOfActividad
      parameters:
        - $ref: '#/components/parameters/idActividad'
      responses:
        '200':
          $ref: '#/components/responses/ExcepcionListaData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Actividad no encontrada
  /recurrencias/{idRecurrencia}:
    get:
      tags:
        - recurrencias
      summary: Devuelve información de una recurrencia por su id
      description: Devuelve información de la recurrencia con id = {idRecurrencia}
      operationId: getRecurrenciaById
      parameters:
        - $ref: '#/components/parameters/idRecurrencia'
      responses:
        '200':
          $ref: '#/components/responses/RecurrenciaData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Recurrencia no encontrada
  /recurrencias/actividades/{idActividad}:
    get:
      tags:
        - recurrencias
        - actividades
      summary: Devuelve la lista de recurrencias asociadas a una actividad
      description: Devuelve la lista de recurrencias de la actividad con id = {idActividad}.
      operationId: getRecurrenciaByActividad
      parameters:
        - $ref: '#/components/parameters/idActividad'
      responses:
        '200':
          $ref: '#/components/responses/RecurrenciaListaData'
        '400':
          description: Id suministrado no válido
        '404':
          description: Actividad no encontrada
...
components:
  parameters:
...
    idUsuario:
      name: idUsuario
      in: path
      description: ID del usuario
      required: true
      schema:
        type: integer
        format: int64
      example: 1
    idActividad:
      name: idActividad
      in: path
      description: ID de la actividad
      required: true
      schema:
        type: integer
        format: int64
      example: 1
    idClase:
      name: idClase
      in: path
      description: ID de la clase
      required: true
      schema:
        type: integer
        format: int64
      example: 1
    idAsignatura:
      name: idAsignatura
      in: path
      description: ID de la asignatura
      required: true
      schema:
        type: integer
        format: int64
      example: 1
    idGrupo:
      name: idGrupo
      in: path
      description: ID del grupo
      required: true
      schema:
        type: integer
        format: int64
      example: 1
...
    idExcepcion:
      name: idExcepcion
      in: path
      description: ID de la excepcion
      required: true
      schema:
        type: integer
        format: int64
      example: 1
    idRecurrencia:
      name: idRecurrencia
      in: path
      description: ID de la recurrencia
      required: true
      schema:
        type: integer
        format: int64
      example: 1
  schemas:
    conId:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Id de la entidad
      required:
        - id
...
    ListaEspacios:
      type: object
      properties:
        espacios:
          type: array
          items:
            $ref: '#/components/schemas/conId'
          description: Array con ids de espacios válidos
          example: [ { "id": 1 } ]
    ListaActividades:
      type: object
      properties:
        actividades:
          type: array
          items:
            $ref: '#/components/schemas/conId'
          description: Array con ids de actividades válidas
          example: [ { "id": 1 } ]
    ActividadTimeData:
      type: object
      properties: 
        fecha_inicio:
          type: string
          description: Fecha de inicio de la actividad y su recursión
          example: "2024-02-14T09:00:00.000"
        fecha_fin:
          type: string
          description: Fecha de fin de la actividad y su recursión
          example: "2024-06-30T18:40:00.000"
        tiempo_inicio:
          type: string
          description: Hora del dia a la que comienza la actividad
        tiempo_fin:
          type: string
          description: Hora del dia a la que finaliza la actividad
        es_todo_el_dia:
          type: string
          enum:
            - Si
            - No
          description: Si es todo el día o no la actividad. Si no lo es se ignoran los campos de tiempo_inicio y tiempo_fin
    ActividadData:
      allOf:
      - $ref: '#/components/schemas/ActividadTimeData'
      - type: object
        properties:
          es_recurrente:
            type: string
            enum:
              - Si
              - No
            description: Si es o no recurrente la actividad
          clase_ids:
              type: array
              items:
                $ref: '#/components/schemas/conId'
              description: Array con ids de clases válidas, para las cuales se realiza la actividad
              example: [ { "id": 1 }, { "id": 3 } ]
    Actividad:
      allOf:
        - $ref: '#/components/schemas/conId'
        - $ref: '#/components/schemas/ActividadData'
    ExcepcionData:
      type: object
      properties:
        esta_reprogramado:
          type: string
          enum:
            - Sí
            - No
          description: Si está o no reprogramada la actividad
        esta_cancelado:
          type: string
          enum:
            - Sí
            - No
          description: Si está o no cancelada la actividad (aunque esté reprogramada, se puede cancelar)
        actividad_id:
          type: integer
          format: int64
          description: Id de la actividad a la que se asocia la excepción
          example: 1
        fecha_inicio_act:
          type: string
          description: fecha de inicio de la actividad (o instancia de actividad) cancelada/reprogramada
          example: '2022-03-01 10:00:00.000'
        fecha_fin_act:
          type: string
          description: fecha de fin de la actividad (o instancia de actividad) cancelada/reprogramada
          example: '2022-03-01 11:40:00.000'
        fecha_inicio_ex:
          type: string
          description: fecha de inicio de la reprogramación de la actividad (o instancia de actividad) reprogramada
          example: '2022-03-05 13:00:00.000'
        fecha_fin_ex:
          type: string
          description: fecha de fin de la reprogramación de la actividad (o instancia de actividad) reprogramada
          example: '2022-03-05 14:40:00.000'
    ListaExcepciones:
      type: object
      properties:
        excepciones:
          type: array
          items:
            $ref: '#/components/schemas/conId'
          description: Array con ids de excepciones válidas
          example: [ { "id": 1 } ]
    ClaseData:
      type: object
      properties:
        asignatura_id:
          type: integer
          format: int64
          description: Id de una asignatura
          example: 1
        grupo_id:
          type: integer
          format: int64
          description: Id de un grupo
          example: 1
    AsignaturaData:
      type: object
      properties:
        nombre:
          type: string
          description: Nombre de la asignatura
          example: "Programación Declarativa"
        siglas:
          type: string
          description: Siglas de la asignatura
          example: "PD"
        departamento:
          type: string 
          description: Departamento que imparte esta asignatura
        periodo:
          type: integer
          enum:
            - 1
            - 2
          description: Periodo / Cuatrimestre en el que se imparte esta asignatura
    GrupoData:
      type: object
      properties:
        curso:
          type: integer
          description: Curso al que pertenece al grupo
          example: 3
        letra:
          type: string
          description: Letra asociada al grupo
          example: "A"
    ListaRecurrencias:
      type: object
      properties:
        recurrencias:
          type: array
          items:
            $ref: '#/components/schemas/conId'
          description: Array con ids de recurrencias válidos
          example: [ { "id": 1 } ]
    RecurrenciaData:
      type: object
      properties:
        tipo_recurrencia:
          type: string
          description: Tipo de recurrencia
          example: "mensual"
        separacion:
          type: integer
          description: Separación en la medida del tipo de recurrencia
          example: 1
        maximo:
          type: integer
          description: Maximo de veces que se repite la recurrencia
          example: 5
        dia_semana:
          type: integer
          description: Dia de la semana en el que ocurre la recurrencia, usado en _semanal
          example: 4
        semana_mes:
          type: integer
          description: Semana del mes en el que ocurre la recurrencia, usado en _mensual
          example: 2
        dia_mes:
          type: integer
          description: Dia del mes en el que ocurre la recurrencia, usado en _mensual
          example: 12
        mes_anio:
          type: integer
          description: Mes del año en el que ocurre la recurrencia, usado en _mensual o _anual
          example: 10
...
  responses:
...
    Actividad:
      description: Datos de una actividad mas su id
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Actividad'
...
    EspacioListaData:
      description: Lista de ids de espacios a devolver a petición de una query
      content:
        application/json:
          schema:
             $ref: '#/components/schemas/ListaEspacios'
    ActividadListaData:
      description: Lista de ids de actividades a devolver a petición de una query
      content:
        application/json:
          schema:
             $ref: '#/components/schemas/ListaActividades'
    ClaseData:
      description: Datos de una clase
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClaseData'
...
    AsignaturaData:
      description: Datos de una asignatura
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AsignaturaData'
    GrupoData:
      description: Datos de un grupo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GrupoData'
...
    ExcepcionData:
      description: Datos de una excepcion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExcepcionData'
    ExcepcionListaData:
      description: Lista de ids de excepciones a devolver a petición de una query
      content:
        application/json:
          schema:
             $ref: '#/components/schemas/ListaExcepciones'
    RecurrenciaListaData:
      description: Lista de ids de recurrencias a devolver a petición de una query
      content:
        application/json:
          schema:
             $ref: '#/components/schemas/ListaRecurrencias'
    RecurrenciaData:
      description: Datos de una recurrencia
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RecurrenciaData'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY